chore: remove lock files (#21)
The yarn.lock and package-lock.json files are out of sync. This means depending on the tool used by the consumer a different set of dependencies will
be installed.

The lock files also result in the template pinning to a specific
version of @elder/elderjs. Currently this is two minor releases behind the latest version. Is that a deliberate decision? It seems that rather than requiring manual updates to this template repo we should defer to the package manager to install the latest version specified by the package.json semver declaration.